Fixed authentication and login
Can now create new user with email and password. New user is stored
in the '/users' node in the firebase database, with email as only
value as of now. Also updates the app state when user is signed in,
which lets all child components use the user data to render. NavBar
conditionally renders and links to loginpage based on this.